refactor: separate responsive logic
